refactor: add indent-blankline and wakatime tools
