+++ nvim/lua/plugins/configs/init.lua
@@ -4,5 +4,6 @@ require 'plugins.configs.lualine'
 require 'plugins.configs.telescope'
 require 'plugins.configs.fzf'
 require 'plugins.configs.bufferline'
-require 'plugins.configs.nvimtreesitter'
-require 'plugins.configs.nvimtreesittercontext'
+require 'plugins.configs.nvim-treesitter'
+require 'plugins.configs.nvim-treesitter-context'
+require 'plugins.configs.indent-blankline'
